+++ src/renderer/src/features/override/external-editor/index.ts
@@ -1,0 +1,4 @@
+export { openInEditor } from './model/openInEditor';
+export { showOverrideFile } from './model/showOverrideFile';
+export { takeFileVersion } from './model/takeFileVersion';
+export { takeOutsideEdits } from './model/takeOutsideEdits';